add syscalls systemtap script
This is a suitable replacement for the existing strace script and is
known to work on Red Hat Enterprise Linux, Oracle Linux, Fedora, and
Ubuntu with a bit of coaxing and gentle prodding. RPM-based
distributions are the most straightforward and tend to work on nearly
every kernel release. Ubuntu frequently breaks.

Basic SystemTap instructions for each distro look something like this:

Red Hat Enterprise Linux:

 - enable the debug RPMs repos using subscription-manager
 - install systemtap, kernel debuginfo, and debug info common

Oracle Linux (9):

 - install systemtap and kernel-uek-devel
 - pull the debug info packages from oss.oracle.com/ol9/debuginfo
 - install the debuginfo and debug info common packages

Fedora Linux:

 - dnf install systemtap-client systemtap-devel systemtap
 - stap-prep

Ubuntu:

 - add the dbgsym release key for ddebs.ubuntu.com
 - add an apt source for ddebs.ubuntu.com
 - update the apt cache
 - install linux-headers-$(uname -r) and linux-image-$(uname -r)-dbgsym

- kevoreilly/CAPEv2#1166